add : added the Topic command (and file)
We can now set a topic to a channel. Everyone can change the topic for
now, will changed it soon. We can also chat on a channel without issue.
We cannot send too big of a message because we don't handle message
being stored in a buffer if it's not send entirely in one go.
So far so good