refactor : topic command
now each use of the topic command has it's own function. The parsing is
still basic but will be handle by the builder. Some features :
We cannot see the topic of a channel if not part of it.
Everyone can change the topic of a channel.
Everyone can clear the topic of a channel
We return the right replies for each 'error' a client might encounter